+++ utils.py
@@ -3,24 +3,100 @@ import typing as t
 import networkx as nx
 import numpy as np
 import pandas as pd
+from sklearn.linear_model import LinearRegression
 
 from definitions import MAX_SCORE_DIFF
 
 
-def duplicate_games(df_games: pd.DataFrame) -> pd.DataFrame:
+def generate_swiss_round(
+    ratings: pd.Series,
+    games: pd.DataFrame,
+    game_ids: t.List[str],
+    round_num: int,
+    pairs_playoffs: t.List[t.Tuple[str, str]] = [],
+):
     """
-    Add duplicates of the games with Team_1 <-> Team_2 and Score_1 <-> Score_2; i.e., each game will be twice
-    in the returned Games Table (some functions are easier to apply on the Games Table in this format).
+    Generate the next group round based on the team ratings.
 
-    :param df_games: Games Table
-    :return: Duplicated Games Table
+    :param ratings: Ratings of non-playoff teams
+    :param games: Games table, DataFrame with columns Round", "Team_1", "Team_2", "Score_1", "Score_2"
+    :param game_ids: Game IDs, acts like the index of the Games tables
+    :param round_num: Round number
+    :param pairs_playoffs: Already determined playoff pairs
+    :return: Games table, DataFrame with columns Round", "Team_1", "Team_2", "Score_1", "Score_2" (scores are empty)
     """
-    df_games_reversed = df_games.rename(
-        columns={"Team_1": "Team_2", "Team_2": "Team_1", "Score_1": "Score_2", "Score_2": "Score_1"}
+    games_next = pd.DataFrame(index=game_ids, columns=["Round", "Team_1", "Team_2", "Score_1", "Score_2"])
+    games_next["Round"] = round_num
+    n_games_playoffs = len(pairs_playoffs)
+    if pairs_playoffs:
+        games_next.loc[game_ids[: n_games_playoffs] , ["Team_1", "Team_2"]] = pairs_playoffs
+    ratings = ratings.loc[~ratings.index.isin([team for pair in pairs_playoffs for team in pair])]
+    games_next.loc[game_ids[n_games_playoffs :], ["Team_1", "Team_2"]] = get_minimal_pairing(ratings, games)
+    return games_next
+
+
+def get_minimal_pairing(ratings: pd.Series, games: pd.DataFrame) -> t.List[t.Tuple[str, str]]:
+    """
+    Generate minimal pairing of the teams based on their ratings
+
+    :param ratings: Ratings
+    :param games: Games table, DataFrame with columns Round", "Team_1", "Team_2", "Score_1", "Score_2"
+    :return: Pairs of next-round opponents
+    """
+    ratings_np = ratings.values
+    teams = ratings.index.tolist()
+    pairing_costs = pd.DataFrame(
+        (ratings_np[:, None] - ratings_np[None, :]) ** 2, index=teams, columns=teams, dtype=float
     )
-    # If Team_Rank_Diff_{...} or Game_Rank_Diff_{...} is present, change the sign in the reversed table
-    df_games_reversed.loc[:, df_games_reversed.columns.str.contains("Rank_Diff")] *= -1
-    return pd.concat([df_games, df_games_reversed]).reset_index(drop=True)
+    if len(teams) % 2 == 1:
+        pairing_costs["pickup"] = 0
+        pairing_costs.loc["pickup"] = 0
+        pairing_costs.loc["pickup", "pickup"] = np.nan
+        ratings = ratings.copy()
+        ratings["pickup"] = -np.inf  # For ordering purposes
+    for team in teams:
+        pairing_costs.loc[team, team] = np.nan
+    for _, rw in games.iterrows():
+        if rw["Team_1"] in teams + ["pickup"] and rw["Team_2"] in teams + ["pickup"]:
+            pairing_costs.loc[rw["Team_1"], rw["Team_2"]] = np.nan
+            pairing_costs.loc[rw["Team_2"], rw["Team_1"]] = np.nan
+    pairing_costs = pairing_costs.stack().reset_index()
+    pairing_costs.columns = ["source", "target", "weight"]
+    pairing_costs["weight"] = 1 + pairing_costs["weight"].max() - pairing_costs["weight"]
+    g = nx.from_pandas_edgelist(pairing_costs, edge_attr="weight")
+    pairs = [p if ratings[p[0]] >= ratings[p[1]] else p[::-1] for p in nx.max_weight_matching(g)]
+    return sorted(pairs, key=lambda x: ratings[x[0]] + ratings[x[1]])[::-1]
+
+
+def calculate_windmill_ratings(
+    games: pd.DataFrame, teams: pd.Series, mean_rating: int = 0, n_round: int = 2
+) -> pd.Series:
+    """
+    Calculate Windmill ratings (linear regression fit of the score differences) based on the given games.
+
+    :param games: Games table, DataFrame with columns Round", "Team_1", "Team_2", "Score_1", "Score_2"
+    :param teams: Teams
+    :param mean_rating: Mean rating
+    :param n_round: Number of decimals to round
+    :return:
+    """
+    games = games.loc[~(games[["Team_1", "Team_2"]] == "pickup").any(axis=1)]
+    games = reorder_games_by_win(games.copy())
+    games["Game_Rank_Diff"] = (games["Score_1"] - games["Score_2"]).clip(upper=MAX_SCORE_DIFF)
+    components = get_graph_components(get_games_graph(games), teams)
+
+    coefficients = pd.Series(0, index=teams, dtype="float64")
+    for i_comp, comp in components.reset_index().groupby("Component"):
+        teams_comp = comp["Team"]
+        df_comp = games.loc[games["Team_1"].isin(teams_comp) | games["Team_2"].isin(teams_comp)].reset_index()
+        df_comp["const"] = 1
+        df_plus = df_comp.pivot(index="index", columns="Team_1", values="const").fillna(0).loc[df_comp["index"]]
+        df_minus = df_comp.pivot(index="index", columns="Team_2", values="const").fillna(0).loc[df_comp["index"]]
+        x = df_plus.subtract(df_minus, fill_value=0)[teams_comp]
+        y = df_comp["Game_Rank_Diff"]
+        lr = LinearRegression(fit_intercept=False).fit(x, y)
+        coefficients[teams_comp] = (lr.coef_ - lr.coef_.mean() + mean_rating).round(n_round)
+    return coefficients.sort_values(ascending=False)
 
 
 def get_summary_of_games(games: pd.DataFrame, ratings: pd.Series, teams: pd.Series) -> pd.DataFrame:
@@ -48,6 +124,22 @@ def get_summary_of_games(games: pd.DataFrame, ratings: pd.Series, teams: pd.Seri
     summary["Score"] = summary.apply(lambda x: f"{x['Goals_for']}-{x['Goals_against']}", axis=1)
     summary = summary[["Rating", "Opponent_Rating", "Record", "Score"]].sort_values(by="Rating", ascending=False)
     return summary
+
+
+def duplicate_games(df_games: pd.DataFrame) -> pd.DataFrame:
+    """
+    Add duplicates of the games with Team_1 <-> Team_2 and Score_1 <-> Score_2; i.e., each game will be twice
+    in the returned Games Table (some functions are easier to apply on the Games Table in this format).
+
+    :param df_games: Games Table
+    :return: Duplicated Games Table
+    """
+    df_games_reversed = df_games.rename(
+        columns={"Team_1": "Team_2", "Team_2": "Team_1", "Score_1": "Score_2", "Score_2": "Score_1"}
+    )
+    # If Team_Rank_Diff_{...} or Game_Rank_Diff_{...} is present, change the sign in the reversed table
+    df_games_reversed.loc[:, df_games_reversed.columns.str.contains("Rank_Diff")] *= -1
+    return pd.concat([df_games, df_games_reversed]).reset_index(drop=True)
 
 
 def get_games_graph(games: pd.DataFrame) -> nx.Graph:
